%pre and %postun always end with 0 now, even on errors. Helps on systems with selinux enabled.
git-svn-id: https://oss.gonicus.de/repositories/gosa/branches/2.5@5447 594d385d-05f5-0310-b6e9-bd551577e9d8

+++ contrib/gosa.spec
@@ -174,12 +174,12 @@ rm -rf %{buildroot}
 /bin/cp /etc/shells /etc/gosa
 
 %pre
-# Cleanup compile dir on updates
-[ -d /var/spool/gosa ] && rm -rf /var/spool/gosa/*
+# Cleanup compile dir on updates, always exit cleanly even on errors
+[ -d /var/spool/gosa ] && rm -rf /var/spool/gosa/* ; exit 0
 
 %postun
 # Remove temporary files, just to be sure
-rm -rf /var/spool/gosa/
+[ -d /var/spool/gosa ] && rm -rf /var/spool/gosa/ ; exit 0
 
 %files
 %defattr(-,%{apacheuser},%{apachegroup})
@@ -221,6 +221,7 @@ rm -rf /var/spool/gosa/
 %changelog
 * Wed Dec 20 2006 Lars Scheiter <lars.scheiter@GONICUS.de> 2.5.7
 - New upstream
+- %pre and %postun always end successfully now, even on errors
 
 * Fri Nov 17 2006 Lars Scheiter <lars.scheiter@GONICUS.de> 2.5.6
 - New upstream
